+++ client/src/App.tsx
@@ -1,4 +1,8 @@
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { useEffect } from 'react';
+import { useNavigate, useLocation, Routes, Route } from 'react-router-dom';
+
+import { useAppDispatch } from './redux/hooks';
+import { setUserId } from './redux/configSlice';
 
 import WelcomePage from './pages/WelcomePage/WelcomePage';
 import LoginPage from './pages/LoginPage/LoginPage';
@@ -13,23 +17,43 @@ import CollectionPage from './pages/CollectionPage/CollectionPage';
 import './App.css';
 
 const App = () => {
+  const navigate = useNavigate();
+  const { pathname } = useLocation();
+  const dispatch = useAppDispatch();
+
+  // todo:
+  // token in cookie
+  // ? how to expire local storage?
+
+  useEffect(() => {
+    const userId = localStorage.getItem('userId');
+    console.log('got it from local', userId);
+
+    if (!userId) {
+      navigate('/welcome');
+    } else {
+      dispatch(setUserId(userId));
+      // todo: use global auth hook
+      // todo: welcome page access with userId - logout
+      navigate(pathname === '/' ? '/home' : `${pathname}`);
+    }
+  }, []);
+
   return (
     <div className="app">
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<WelcomePage />} />
-          <Route path="/login" element={<LoginPage />} />
-          <Route path="/signup" element={<SignupPage />} />
+      <Routes>
+        <Route path="/welcome" element={<WelcomePage />} />
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/signup" element={<SignupPage />} />
 
-          <Route path="/" element={<NavOutlet />}>
-            <Route path="/home" element={<HomePage />} />
-            <Route path="/profile" element={<ProfilePage />} />
-          </Route>
-          <Route path="/config/:type" element={<ConfigPage />} />
-          <Route path="/wishlist" element={<WishlistPage />} />
-          <Route path="/collection" element={<CollectionPage />} />
-        </Routes>
-      </BrowserRouter>
+        <Route path="/" element={<NavOutlet />}>
+          <Route path="/home" element={<HomePage />} />
+          <Route path="/profile" element={<ProfilePage />} />
+        </Route>
+        <Route path="/config/:type" element={<ConfigPage />} />
+        <Route path="/wishlist" element={<WishlistPage />} />
+        <Route path="/collection" element={<CollectionPage />} />
+      </Routes>
     </div>
   );
 };
