+++ client/src/pages/SignupPage/SignupPage.tsx
@@ -58,8 +58,8 @@ const SignupPage = () => {
     signup(user)
       .unwrap()
       .then((res) => {
-        console.log(res);
         const { userId } = res;
+        localStorage.setItem('userId', userId);
         dispatch(setUserId(userId));
         navigate('/home');
       });
